requirements.txt
mysql-connector-python==8.0.28
flask==2.0.3
flask-restful==0.3.9